refactor: move LandingPage
