+++ client/src/App.tsx
@@ -1,5 +1,5 @@
 import type { Component } from "solid-js";
-import LandingPage from "./app/pages/LandingPage/LandingPage";
+import LandingPage from "./app/pages/LandingPage";
 
 const App: Component = () => <LandingPage />;
